refactor: organize orderbook queries
